+++ app/javascript/controllers/trips_controller.js
@@ -7,10 +7,8 @@ import BaseController from "./base_controller"
 import L from "leaflet"
 import { createAllMapLayers } from "../maps/layers"
 import { createPopupContent } from "../maps/popups"
-import {
-  fetchAndDisplayPhotos,
-  showFlashMessage
-} from '../maps/helpers';
+import { showFlashMessage } from '../maps/helpers';
+import { fetchAndDisplayPhotos } from '../maps/photos';
 
 export default class extends BaseController {
   static targets = ["container", "startedAt", "endedAt"]